Improve test docstrings, remove non-ASCII symbols and rename helper file
