+++ tests/test_inference.py
@@ -1,3 +1,14 @@
+"""Device-agnostic inference plumbing: placement, AMP, auto-batching, async copies.
+
+CUDA, XPU, and MPS share one inference path (CPU is the only early-return
+exception).
+
+CUDA and XPU expose async copy primitives, while MPS does not.
+
+Other devices backends should still work, but will fall back to synchronous
+copies with a warning if they don't expose all the required primitives.
+"""
+
 import warnings
 from collections import OrderedDict
 from unittest.mock import MagicMock
@@ -28,6 +39,7 @@ from tabicl._model.inference import AsyncCopyManager, InferenceManager, devices_
     ],
 )
 def test_devices_match(left, right, expected):
+    """Unindexed devices match index 0; different indices or backends do not."""
     assert devices_match(torch.device(left), torch.device(right)) is expected
 
 
@@ -54,13 +66,14 @@ def test_is_tensor_on_exe_device_accepts_unindexed_and_indexed(device_backend):
     on_cpu = torch.zeros(1, device="cpu")
     assert mgr._is_tensor_on_exe_device(on_device) is True
     assert mgr._is_tensor_on_exe_device(on_cpu) is (device_backend == "cpu")
-    # Avoid a redundant host→device copy when the tensor is already placed correctly.
+    # Avoid a redundant host -> device copy when the tensor is already placed correctly.
     moved = mgr._to_exe_device(on_device)
     assert moved is on_device
 
 
 @pytest.mark.parametrize("device_backend", ["cuda", "xpu", "mps"])
 def test_get_available_gpu_memory_non_zero_for_available_backends(device_backend):
+    """Check that device memory inspection works and always returns positive values."""
     backend_api = getattr(torch, device_backend, None)
     if (
         backend_api is None
@@ -88,30 +101,17 @@ class _FakeNoAsyncBackend:
         return True
 
 
-@pytest.mark.parametrize("device_backend", ["cuda", "xpu"])
-def test_async_copy_manager_uses_backend_async_primitives(device_backend):
-    backend_api = getattr(torch, device_backend, None)
-    if (
-        backend_api is None
-        or not callable(getattr(backend_api, "is_available", None))
-        or not backend_api.is_available()
-    ):
-        pytest.skip(f"{device_backend} backend is not available on this host")
+@pytest.mark.parametrize("device_type", ["cuda", "xpu"])
+def test_async_copy_manager_uses_backend_async_primitives(device_type):
+    """Check that theasync copy manager uses CUDA/XPU Stream/Event primitives.
 
-    required_primitives = ("Stream", "Event", "current_stream", "stream")
-    if any(
-        not callable(getattr(backend_api, primitive, None))
-        for primitive in required_primitives
-    ):
-        pytest.skip(
-            f"{device_backend} backend does not expose full async stream/event primitives"
-        )
-
-    device = torch.device(device_backend)
+    CUDA and XPU device backends are expected to always expose Stream/Event primitives.
+    """
+    device = torch.device(device_type)
     manager = AsyncCopyManager(device=device, max_pending=2)
     if manager._copy_stream is None:
         pytest.skip(
-            f"{device_backend} async stream setup is not supported in this runtime"
+            f"{device_type} async stream setup is not supported in this runtime"
         )
 
     src = torch.ones((4,), dtype=torch.float32, device=device)
@@ -128,10 +128,13 @@ def test_async_copy_manager_uses_backend_async_primitives(device_backend):
 
     assert len(manager._pending) == 0
     assert torch.equal(dst, src.cpu())
-    assert manager.get_bytes_written() > 0.0
+    assert manager.get_bytes_written() == pytest.approx(
+        src.numel() * src.element_size() / (1024 * 1024)
+    )
 
 
 def test_async_copy_manager_falls_back_to_sync_without_async_primitives(monkeypatch):
+    """Other unknown backends without Stream/Event copy synchronously and warn."""
     backend_name = "fakenoasync"
     monkeypatch.setattr(torch, backend_name, _FakeNoAsyncBackend, raising=False)
 
@@ -147,7 +150,9 @@ def test_async_copy_manager_falls_back_to_sync_without_async_primitives(monkeypa
     # No async primitives: direct sync copy, no pending events.
     assert len(manager._pending) == 0
     assert torch.equal(dst, src)
-    assert manager.get_bytes_written() > 0.0
+    assert manager.get_bytes_written() == pytest.approx(
+        src.numel() * src.element_size() / (1024 * 1024)
+    )
 
 
 @pytest.mark.skipif(not torch.backends.mps.is_available(), reason="MPS not available")
@@ -266,6 +271,12 @@ def test_mps_auto_batch_uses_memory_estimate(monkeypatch):
 
 @pytest.mark.skipif(not torch.backends.mps.is_available(), reason="MPS not available")
 def test_mps_memory_estimate_uses_recommended_minus_current(monkeypatch):
+    """The MPS memory inspection API is specific compared to CUDA/XPU.
+
+    In particular, it does not expose mem_get_info, and the available memory is
+    therefore estimated as recommended_max_memory minus
+    current_allocated_memory.
+    """
     mgr = InferenceManager(enc_name="tf_col", out_dim=4)
     mgr.configure(device="mps")
 
@@ -278,5 +289,5 @@ def test_mps_memory_estimate_uses_recommended_minus_current(monkeypatch):
     fake_api.empty_cache = MagicMock()
 
     monkeypatch.setattr(mgr, "_get_device_backend_api", lambda: fake_api)
-    available_mb = mgr.get_available_gpu_memory()
-    assert available_mb == pytest.approx(6 * 1024.0)
+    available_mib = mgr.get_available_gpu_memory()
+    assert available_mib == pytest.approx(6 * 1024.0)
