+++ tests/torch_devices_helpers.py
